ADD: protecting video routes

+++ controllers/videoController.js
@@ -1,9 +1,11 @@
 import routes from "../routes";
 import Video from "../models/Video";
 
+// Home
+
 export const home = async (req, res) => {
   try {
-    const videos = await Video.find({}).sort({_id:-1});
+    const videos = await Video.find({}).sort({ _id: -1 });
     res.render("home", { pageTitle: "Home", videos });
   } catch (error) {
     console.log(error);
@@ -11,20 +13,24 @@ export const home = async (req, res) => {
   }
 };
 
-export const search = async(req, res) => {
+// Search
+
+export const search = async (req, res) => {
   const {
     query: { term: searchingBy }
   } = req;
-  let videos = []
+  let videos = [];
   try {
     videos = await Video.find({
-      title:{$regex:searchingBy,$options:"i"}
+      title: { $regex: searchingBy, $options: "i" }
     });
   } catch (error) {
-    console.log(error)
+    console.log(error);
   }
-  res.render("search",{pageTitle:"Search",searchingBy, videos})
+  res.render("search", { pageTitle: "Search", searchingBy, videos });
 };
+
+// Upload
 
 export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload" });
@@ -37,22 +43,29 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
+
+// Video Detail
 
 export const videoDetail = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
   }
 };
+
+// Edit Video
 
 export const getEditVideo = async (req, res) => {
   const {
@@ -60,7 +73,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -79,12 +96,19 @@ export const postEditVideo = async (req, res) => {
   }
 };
 
+// Delete Video
+
 export const deleteVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
